+++ src/wtime.hpp
@@ -1,3 +1,3 @@
 #pragma once
 // current time in seconds
-double wtime(void);
+float wtime(void);
